--- tilelang/ir/__init__.py
@@ -1,9 +1,0 @@
-from .fill import Fill  # noqa: F401
-from .atomic_add import AtomicAdd  # noqa: F401
-from .copy import Copy, Conv2DIm2ColOp  # noqa: F401
-from .gemm import Gemm  # noqa: F401
-from .gemm_sp import GemmSP  # noqa: F401
-from .finalize_reducer import FinalizeReducerOp  # noqa: F401
-from .parallel import ParallelOp  # noqa: F401
-from .reduce import ReduceOp, CumSumOp  # noqa: F401
-from .region import RegionOp  # noqa: F401
