--- tilelang/ir/atomic_add.py
@@ -1,8 +1,0 @@
-from tilelang import tvm as tvm
-from tvm.ir import Node, Scriptable
-import tvm.ffi
-
-
-@tvm.ffi.register_object("tl.AtomicAdd")
-class AtomicAdd(Node, Scriptable):
-    ...
