--- tilelang/ir/parallel.py
@@ -1,8 +1,0 @@
-from tilelang import tvm as tvm
-from tvm.ir import Node, Scriptable
-import tvm.ffi
-
-
-@tvm.ffi.register_object("tl.ParallelOp")
-class ParallelOp(Node, Scriptable):
-    ...
